Update regex to remove single-space delim support. Add more spec tests. (cherry picked from commit c71bf46713774d33ce9d09a756992c2e201198a9)
@clintslee 's work
There was an issue with the regex not matching the string ", ", which is fixed.
This commit is cherrypicked due to a messy merge from master in the feature branch.
Additionally, there was a missing schema.rb change that needed to be migrated in,
if this problem keeps occuring, run `rake db:migrate RAILS_ENV=test`.