Add link to new order view. Fixes #61
Somewhere in the devise swapover we lost the logic that directs
non-admin users to the new order page by default, and there wasn't a
link to it anywhere on the orders list. This adds a "new order" button
visible to non-admin users, and is mostly only intended as a stop-gap
for Patrick to do his user demo. We still need to implement a cohesive
global nav (#31).