Remove Request model
Having nested requests / multiple supplies inside an order leads to a
*lot* of extra complication:
- growing form input for extra supplies
- having to use nested attributes and the extra validation problems
- difficult duplicate checks
- having to worry about partial fulfillment of an order
- diverging web and sms flows (since SMS always only requests one
  supply)

Really, once an order is in the system, we typically want to track each
supply separately. This commit removes the request model and treats
orders as the user<->supply join model. If we ever need to group orders
into one placement unit, we can add an extra id column later.