Allow "request for". #100, 97
This wires up the "place a request on behalf of ..." functionality that
PCMOs and admins need.

@jasnow - I can't run cukes on my machine because of something Headless
related ("Xvfb not found on your system (Headless::Exception)"), so I've
stubbed out the .feature file. I'm going to assign these to you to
finish up the feature testing (and to kick back my way when they're
broken). Can you also help me understand what I need to do to get the
cukes running on OSX?