Replace client-side filter with active country in session #295, #304
Admins now set an active country and then manage orders as a PCMO from
that country would. This should solve most of the "forgetful" complaints
we've had about admin activity, along with allowing us to sort and
paginate everything easily and consistently server-side.

The filter was the most substantial bit of javascript, so I went ahead
and moved the other bits to where they are used. It's not as cleanly
extracted, but for infrequently-developed codebases like this, the
increased code locality is a win.