Raise maximum length of response text
There are some existing long responses in the system that we will need to save again (for confirmation), and we don't want to invalidate those. But we do need to make sure new records fit in the DB.

The UI should already constrain inputs to 160 characters, so this is mostly for browsers where that doesn't work (looking at you, IE).